requirements.txt
aiohttp>=3.8.4
redis>=4.6.0
openai>=0.27.0
python-dotenv>=1.0.0
